medical_appointment/requirements.txt
streamlit
joblib
pandas
numpy
scikit-learn
pillow
xgboost
cloudpickle